requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.1.1
Flask==1.1.2
gitdb==4.0.5
GitPython==3.1.8
google-api-core==1.22.2
google-auth==1.21.2
google-cloud==0.34.0
google-cloud-automl==2.0.0
google-cloud-core==1.4.1
google-cloud-storage==1.31.0
google-cloud-vision==1.0.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
idna==2.10
image==1.5.32
itsdangerous==1.1.0
Jinja2==2.11.2
libcst==0.3.10
MarkupSafe==1.1.1
mypy-extensions==0.4.3
Pillow==7.2.0
proto-plus==1.9.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
smmap==3.0.4
sqlparse==0.3.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
Werkzeug==1.0.1
